requirements.txt
python-dotenv
streamlit
google-cloud-aiplatform[adk,agent_engines]
streamlit-local-storage